+++ docs/changelog.md
@@ -275,6 +275,12 @@
 - make `StringU*` use `NonMax*` for the length.
 - remove `StringU32`, `StringUSize`.
 
+#### text::unicode::scalar
+- update `Char`:
+  - improve documentation.
+  - add methods: `ceil_utf8_boundary`, `floor_utf8_boundary`, `is_utf8_continuation`.
+  - add missing method for ref array: `to_char_unchecked`.
+
 ### ui::event
 - new types: `EventButtons`, `EventTagSet`, `EventWheelUnit`.
 - new alias `EventKindTimed`.